Made some QRCodeInfo and Comment Class. Also made QRCodeInfo Activity
This is so that there is a way to access the comment class. QRCodeInfo will not have many functionalities yet, but it will prolly just be leading to other activities, in which I can add later on.

QR code Selected is just a place holder at the moment. I still need to add a back button.

Main Goal is to Get comment class working properly then anything of the activities that surrounds it made and leads to comment activity and maybe some of its functionality working.

The design is based on the project part 2 Miro "Comment" design