Just got complicated. Need to make sure that I can add this to the Firebase Database.
Need to be able to get the comment that is being added, and add it to the firebase, then be able to retract that from the firebase and view it on the activity, so everyone can see the comment. Also need to make the comment section have the attributes of who wrote it? when? and make it place in order of the first one being the oldest and the last one being the youngest comment.