Almost done implementing User Story implementation.
Need: Firestore Database management.
Need to figure out how the comments of this qr code and load it in from the database and then add some files into the database as well.

Because my thinking is that each QR code has a set of Comments, and so we just don't have a big data set of comments that needs to be sorted by QrCode hash value.