Fixed the MainActivity by taking away the Testing comment methods and buttons
All that needs to be done now is the getting the username, and the Specific QR code picture, and then
showing it on the top of the Comment Activity. Also need to make the fragment that calls the CommentActivity.
The CommentActivity needs to get the received message from the Fragment thought, specifically the QRCodeID.