Profile, ProfileController, DatabaseController, & LeaderBoardController Update
A profile now stores it's points, not it's ranking
DatabaseController has had a parameter removed from it's constructor, it no longer needs a FireStore instance passed to it and now gets the instance itself.
Profile Controller has been created & javadoc'ed
LeaderboardController has been created, but is poorly designed. Currently a leaderboard ranking is not stored in Firebase and the creation of the leaderboard is done every time the controller is initialized. Lots of unnecessary work that could be saved remotely.

Everything is untested lol.