Removed MyStatsActivity, removed unecessary strings in the resources
MyStatsActivity had no new information that could not be found in ViewProfile or LeaderboardActivity.
.java and .xml file deleted, activity removed from manifest file.
ViewProfileActivity updated to also show user's ranking
Also removed 90% of the strings in the string resources file, because most were not used anywhere.
Kept all of the resource strings that were actually used someplace in the app