Refactoring methods in Camera Activity & Controller. Incomplete and bug prone.
Changed camera controller to have a method to take a photo and return a Bitmap via ListeningList<>, and have a separate method to scan a QRCode from a Bitmap.
There is a bug in the new method that turns an Image object into a Bitmap object. Need to figure out what type of Image object the Camera Scan returns.
Theoretically, once this is fixed, the Camera Activity should now be able to save the image taken to the database as well.
Bugfixing to follow