Logout confirm popup
AlertDialog fragment popup appears on hitting logout button.
There is a button to cancel logout, confirm the logout, and open the GenerateActivity.
Popup can not generate a QRCode on it's own. Could be a neat feature for polish, but this small bit of polish will take a considerable amount of work to implement. This setup is currently quite simple, not much new code was added.